refactor: general folder structure
