+++ lib/decorators/.eslintrc
