--- lib/decorators/httpCode/index.ts
@@ -1,1 +1,0 @@
-export * from './httpCode.decorator';
