--- lib/decorators/httpMethod/index.ts
@@ -1,1 +1,0 @@
-export * from './httpMethod.decorators';
