+++ lib/decorators/index.ts
@@ -1,4 +1,4 @@
-export * from './httpCode';
-export * from './httpMethod';
-export * from './parameter';
-export * from './setHeader';
+export * from './httpCode.decorator';
+export * from './httpMethod.decorators';
+export * from './parameter.decorators';
+export * from './setHeader.decorator';
