--- lib/decorators/parameter/index.ts
@@ -1,1 +1,0 @@
-export * from './parameter.decorators';
