--- lib/decorators/setHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './setHeader.decorator';
