+++ lib/exceptions/NotFoundException.ts
@@ -1,4 +1,4 @@
-import { HttpException } from '../HttpException';
+import { HttpException } from './HttpException';
 
 export class NotFoundException extends HttpException {
   public name = 'NotFoundException';
