--- lib/exceptions/common/index.ts
@@ -1,5 +1,0 @@
-export * from './BadRequestException';
-export * from './InternalServerErrorException';
-export * from './NotFoundException';
-export * from './UnauthorizedException';
-export * from './UnprocessableEntityException';
