+++ lib/exceptions/index.ts
@@ -1,2 +1,6 @@
 export * from './HttpException';
-export * from './common';
+export * from './BadRequestException';
+export * from './InternalServerErrorException';
+export * from './NotFoundException';
+export * from './UnauthorizedException';
+export * from './UnprocessableEntityException';
