+++ lib/pipes/index.ts
@@ -1,5 +1,2 @@
-export * from './validation.pipe';
-export * from './parseBoolean';
-export * from './parseNumber';
-export * from './validateEnum';
-export * from './parseDate';
+export * from './validators';
+export * from './parsers';
