--- lib/pipes/parseBoolean/index.ts
@@ -1,1 +1,0 @@
-export * from './parseBoolean.pipe';
