--- lib/pipes/parseDate/index.ts
@@ -1,1 +1,0 @@
-export * from './parseDate.pipe';
