--- lib/pipes/parseNumber/index.ts
@@ -1,1 +1,0 @@
-export * from './parseNumber.pipe';
