+++ lib/pipes/parsers/parseBoolean.pipe.ts
