+++ lib/pipes/validators/index.ts
@@ -1,1 +1,2 @@
+export * from './validation.pipe';
 export * from './validateEnum.pipe';
