+++ lib/pipes/validators/validateEnum.pipe.ts
