+++ lib/pipes/validators/validation.pipe.ts
@@ -1,7 +1,7 @@
 import type { ClassTransformOptions } from 'class-transformer';
 import type { ValidatorOptions } from 'class-validator';
-import { validateObject } from '../internals/classValidator';
-import type { ParameterPipe, PipeMetadata } from './ParameterPipe';
+import { validateObject } from '../../internals/classValidator';
+import type { ParameterPipe, PipeMetadata } from '../ParameterPipe';
 
 export interface ValidationPipeOptions extends ValidatorOptions {
   transformOptions?: ClassTransformOptions;
